# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PhotosDen
  version: 1.0.0
  description: Secure, serverless personal photo storage on AWS.
servers:
  - url: https://g2l2bkihl6.execute-api.ap-south-1.amazonaws.com/dev
    description: Development Environment
# Security is now handled in Lambda for better CORS control
# security:
#   - BearerAuth: []
paths:
...
  /albums/{albumId}:
    get:
      summary: Get album by ID
      tags: [Albums]
      security:
        - BearerAuth: []
      x-amazon-apigateway-integration:
        httpMethod: POST
        type: aws_proxy
        uri: ${lambda_invoke_arn}
        payloadFormatVersion: "2.0"
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Album details
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/X-Correlation-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album" 
    delete:
      summary: Soft delete album
      tags: [Albums]
      security:
        - BearerAuth: []
      x-amazon-apigateway-integration:
        httpMethod: POST
        type: aws_proxy
        uri: ${lambda_invoke_arn}
        payloadFormatVersion: "2.0"
        responses:
          default:
            statusCode: "204"
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Album soft deleted
          headers:
            X-Correlation-Id:
              $ref: "#/components/headers/X-Correlation-Id"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito User Pool Access Token
      x-amazon-apigateway-authorizer:
        type: "cognito_user_pools"
        providerArns:
          - "${cognito_user_pool_arn}"
      x-amazon-apigateway-authtype: "cognito_user_pools"
  headers:
    X-Correlation-Id:
      description: Unique ID for request tracing
      schema:
        type: string
...
  schemas:
...
    Album:
      type: object
      required: [title, isDeleted]
      properties:
        title:
          type: string
        description:
          type: string
        isDeleted:
          type: boolean
        deletedAt:
          type: string
          format: date-time
        purgeAt:
          type: integer
          description: Epoch seconds (TTL)
...